Edited Strip SAR tradespace and Added ScanSAR operation
Altered StripSar function, so that it is correct, what was found was that the SNR was not the limiting factor ~15dB when requirement is 5dB. This means the integration time can be reduced so ScanSAR script was made to increase effective FoV

When using Herrick-Gibbs or Full state observation orbit determination, ScanSAR can provide a larger FOV at a given range.

This makes Herrick-Gibbs feasible for up to 50km range if 2 satellites are used in series with overlapping FoV. This increases the effective FoV to the Herrick-Gibbs method.

If full state observation is possible, this also enables a wider FOV than StripSAR.

Overall, this ScanSAR is better and should be used no matter the orbit determination method.

Need to see if the number of subswaths is viable.